requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-appconf==1.0.2
django-js-asset==1.0.0
django-mediumeditor==1.0.0
django-meta==1.4.1
django-widget-tweaks==1.4.1
idna==2.6
Pillow==5.0.0
pycryptodomex==3.4.7
PyJWT==1.3.0
pytz==2018.3
requests==2.18.4
six==1.11.0
sorl-thumbnail==12.4.1
urllib3==1.22
